# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Erro de domínio das rotas `/auth`, `/me` e `/admin` (ADR-008).",
        "example": {
          "error": "user_not_found",
          "message": "Usuário não encontrado."
        },
        "properties": {
          "error": {
            "description": "Slug estável do erro, ex.: `user_not_found`.",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "description": "Mensagem legível, em português.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ExecutionDetail": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/ExecutionAnalysisRef"
          },
          "cached": {
            "title": "Cached",
            "type": "boolean"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "executed_at": {
            "format": "date-time",
            "title": "Executed At",
            "type": "string"
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "result_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Size Bytes"
          },
          "rows_affected": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Affected"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionUserRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "analysis",
          "user",
          "status",
          "error_code",
          "cached",
          "parameters",
          "execution_time_ms",
          "rows_affected",
          "result_size_bytes",
          "executed_at",
          "error_message"
        ],
        "title": "ExecutionDetail",
        "type": "object"
      },
      "ExecutionStats": {
        "properties": {
          "by_error_code": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Error Code",
            "type": "object"
          },
          "by_status": {
            "$ref": "#/components/schemas/StatusCounts"
          },
          "cache": {
            "$ref": "#/components/schemas/CacheStats"
          },
          "execution_time_ms": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TimeStats"
            },
            "title": "Execution Time Ms",
            "type": "object"
          },
          "period": {
            "additionalProperties": {
              "format": "date-time",
              "type": "string"
            },
            "title": "Period",
            "type": "object"
          },
          "top_analyses": {
            "items": {
              "$ref": "#/components/schemas/ExecutionTopAnalysis"
            },
            "title": "Top Analyses",
            "type": "array"
          },
          "top_users": {
            "items": {
              "$ref": "#/components/schemas/ExecutionTopUser"
            },
            "title": "Top Users",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "period",
          "total",
          "by_status",
          "by_error_code",
          "cache",
          "execution_time_ms",
          "top_analyses",
          "top_users"
        ],
        "title": "ExecutionStats",
        "type": "object"
      },
      "ExecutionSummary": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/ExecutionAnalysisRef"
          },
          "cached": {
            "title": "Cached",
            "type": "boolean"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "executed_at": {
            "format": "date-time",
            "title": "Executed At",
            "type": "string"
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "result_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Size Bytes"
          },
          "rows_affected": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Affected"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionUserRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "analysis",
          "user",
          "status",
          "error_code",
          "cached",
          "parameters",
          "execution_time_ms",
          "rows_affected",
          "result_size_bytes",
          "executed_at"
        ],
        "title": "ExecutionSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ExecutionSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExecutionSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ExecutionSummary]",
        "type": "object"
      },
...
      "StatusCounts": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "integer"
          },
          "success": {
            "title": "Success",
            "type": "integer"
          },
          "timeout": {
            "title": "Timeout",
            "type": "integer"
          },
          "volume_exceeded": {
            "title": "Volume Exceeded",
            "type": "integer"
          }
        },
        "required": [
          "success",
          "volume_exceeded",
          "error",
          "timeout"
        ],
        "title": "StatusCounts",
        "type": "object"
      },
...
      "ValidationErrorBody": {
        "description": "422 de validação de corpo/parâmetros no formato padrão do FastAPI (não é o slug `error`/`message`).",
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ValidationErrorBody",
        "type": "object"
      },
      "ValidationErrorItem": {
        "properties": {
          "loc": {
            "description": "Caminho do campo inválido, ex.: `[\"body\", \"email\"]`.",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "description": "Descrição do problema.",
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "description": "Tipo do erro de validação (pydantic).",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationErrorItem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerToken": {
        "description": "Token opaco emitido por `POST /auth/token`. Cole apenas o token (o Swagger acrescenta o prefixo `Bearer`).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API HTTP do servidor de análise de dados via MCP: emissão de tokens (`/auth`), dados do próprio\nusuário (`/me`), administração (`/admin/*`) e `/health`.\n\n**Autenticação.** Emita um token em `POST /auth/token` (e-mail + senha), clique em **Authorize**\ne cole só o token. `/me` exige um token válido; `/admin/*` exige também administrador.\n\n**Erros.** As rotas de `/auth`, `/me` e `/admin` respondem erros de domínio no formato\n`{\"error\": \"<slug>\", \"message\": \"...\"}`. Erros de validação do corpo ou dos parâmetros vêm no\nformato padrão do FastAPI (`422`, `{\"detail\": [...]}`).\n\n**O `/mcp` não está neste documento.** O endpoint MCP (Streamable HTTP/JSON-RPC) é descrito em\n`docs/MCP.md`.\n",
    "title": "Análise de Dados Genérica com MCP — API HTTP",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/analyses/{analysis_id}/executions": {
      "get": {
        "description": "Atalho de `GET /admin/executions` já filtrado pela análise (aceita também as análises inativas). Mesmos filtros, sem `analysis_id`.",
        "operationId": "list_analysis_executions_admin_analyses__analysis_id__executions_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "success",
                    "volume_exceeded",
                    "error",
                    "timeout"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "error_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ANALYSIS_NOT_FOUND",
                    "INVALID_PARAMETERS",
                    "INVALID_ANALYSIS_CONFIG",
                    "DATA_SOURCE_UNAVAILABLE",
                    "QUERY_TIMEOUT",
                    "QUERY_FAILED",
                    "INTERNAL_ERROR"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Code"
            }
          },
          {
            "in": "query",
            "name": "cached",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cached"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "min_time_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Time Ms"
            }
          },
          {
            "in": "query",
            "name": "parameters",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parameters"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExecutionSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `analysis_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 List Analysis Executions Admin Analyses  Analysis Id  Executions Get"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_period`, `invalid_parameters_filter`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Histórico da análise",
        "tags": [
          "admin-analyses"
        ]
      }
    },
...
    "/admin/executions": {
      "get": {
        "description": "Lista paginada do histórico, mais recentes primeiro, com filtros por análise, usuário, status, `error_code`, cache, período (`from`/`to`; padrão: últimos 7 dias, máximo 366), tempo mínimo e `parameters` (JSON, igualdade de chave/valor).",
        "operationId": "list_executions_admin_executions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "analysis_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Analysis Id"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "success",
                    "volume_exceeded",
                    "error",
                    "timeout"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "error_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ANALYSIS_NOT_FOUND",
                    "INVALID_PARAMETERS",
                    "INVALID_ANALYSIS_CONFIG",
                    "DATA_SOURCE_UNAVAILABLE",
                    "QUERY_TIMEOUT",
                    "QUERY_FAILED",
                    "INTERNAL_ERROR"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Code"
            }
          },
          {
            "in": "query",
            "name": "cached",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cached"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "min_time_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Time Ms"
            }
          },
          {
            "in": "query",
            "name": "parameters",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parameters"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExecutionSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 List Executions Admin Executions Get"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_period`, `invalid_parameters_filter`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Listar execuções",
        "tags": [
          "admin-executions"
        ]
      }
    },
    "/admin/executions/stats": {
      "get": {
        "description": "Totais por status e `error_code`, taxa de cache hit, tempo (cache hit × miss: média, p95, máximo) e rankings de análises e usuários (`top`). Janela máxima definida pelo servidor.",
        "operationId": "execution_stats_admin_executions_stats_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "analysis_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Analysis Id"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "top",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Top",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Execution Stats Admin Executions Stats Get"
                }
              }
            },
            "description": "Entidade não processável. Slugs: `invalid_period`. Também pode vir no formato padrão do FastAPI (`{\"detail\": [...]}`) quando o corpo ou os parâmetros não passam na validação."
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Estatísticas de execuções",
        "tags": [
          "admin-executions"
        ]
      }
    },
    "/admin/executions/{execution_id}": {
      "get": {
        "description": "Devolve a execução com `error_message`.",
        "operationId": "get_execution_admin_executions__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não autenticado. Slugs: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sem permissão. Slugs: `forbidden`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não encontrado. Slugs: `execution_not_found`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerToken": []
          }
        ],
        "summary": "Detalhar execução",
        "tags": [
          "admin-executions"
        ]
      }
    },
...
  },
...
}